+++ lib/index.js
@@ -23,7 +23,7 @@ module.exports = (opt = {}) => {
 
     if (config.strictMode && !store.getTotal()) {
       debug(`Database is empty and stricMode is enable. ${ip} is rejected`)
-      next('Tor Database is empty')
+      return next('Tor Database is empty')
     }
 
     req.isTorUser = isExitNode(ip)
